// src/services/auth/interface.ts
export enum AuthErrors {
  EMAIL = 'Введите Email',
  NAME = 'Введите Имя',
  LASTNAME = 'Введите Фамилию',
  PHONE_NUMBER = 'Введите номер телефона',
  PASSWORD = 'Введите Пароль',
  INVALID_OTP = 'Неверный код подтверждения',
  UNAUTHORIZED = 'Неверный логин или пароль',
}